Add multinode CI job
Packstack has traditionally allowed multinode deployments with
1 controller, one or more network nodes and several compute nodes,
but this has never been properly tested in our CI.

This job enables an option to use a 2-node architecture, with one
controller and one compute node. It adds a job running scenario002,
but enables scenario001 and scenario003, so they can be used on
demand if needed.

Change-Id: Ide96a232c902370f3ec25ed54f2b272cc1757dad