Temporary switch to ovs in scenario 001, fix gate
Until https://bugs.launchpad.net/packstack/+bug/1771500
is fixed, temporary switch to ovs in scenario 001.

Also, set inject_partition to the default value of -2 in nova.conf.
We are having problems in the gate in OVH nodes (see [1]), and
doing this prevents us from using the problematic libguestfs code.

[1] - https://bugs.launchpad.net/packstack/+bug/1701032

Change-Id: Ifed85df073104684cd415792785acd12e2f89aa4
Related-Bug: #1771500
Co-Authored-By: Javier Peña <jpena@redhat.com>